separate answerTypes and answerCreators
